Revert "manage: For isfloating/oldstate check/set, ensure trans client actually exists"
This reverts commit bece862a0fc4fc18ef9065b18cd28e2032d0d975.

With this commit, yes, many windows like with SDL are tiled, but so are
windows that I don't want tiled, like GPG pinentry.

I'll wait to apply this when these issues are addressed, if ever. In the
meantime I've found that with SDL programs, passing the flag
SDL_WINDOW_RESIZABLE to SDL_CreateWindow() will make a window tilable
just the same.